requirements.txt
# Core dependencies
google-generativeai>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0

# Utilities
colorama>=0.4.6
rich>=13.0.0
